Supports removing nodes [instable]
Possibile to exit a running Constellation instance by calling
constellation.done() with the property allowLeaving set to "true". The
node will notify all other connected nodes that it is leaving and sleep
for 1 second before actually leaving the IBIS pool and Constellation.

This needs to be extended with acknowledgments and fault-tolerance to
ensure that no messages/activities are lost when exiting.